+++ src/Mod/Fem/App/FemConstraintBearing.cpp
@@ -50,9 +50,9 @@ ConstraintBearing::ConstraintBearing()
     ADD_PROPERTY(AxialFree,(0));
     ADD_PROPERTY(Radius,(0.0));
     ADD_PROPERTY(Height,(0.0));
-    ADD_PROPERTY_TYPE(BasePoint,(Base::Vector3f(0,0,0)),"ConstraintBearing",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
+    ADD_PROPERTY_TYPE(BasePoint,(Base::Vector3d(0,0,0)),"ConstraintBearing",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
                       "Base point of cylindrical bearing seat");
-    ADD_PROPERTY_TYPE(Axis,(Base::Vector3f(0,1,0)),"ConstraintBearing",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
+    ADD_PROPERTY_TYPE(Axis,(Base::Vector3d(0,1,0)),"ConstraintBearing",App::PropertyType(App::Prop_ReadOnly|App::Prop_Output),
                       "Axis of bearing seat");
 }
 
@@ -71,7 +71,7 @@ void ConstraintBearing::onChanged(const App::Property* prop)
     if (prop == &References) {
         // Find data of cylinder
         float radius, height;
-        Base::Vector3f base, axis;
+        Base::Vector3d base, axis;
         if (!getCylinder(radius, height, base, axis))
             return;
         Radius.setValue(radius);
@@ -107,7 +107,7 @@ void ConstraintBearing::onChanged(const App::Property* prop)
         }
 
         float radius, height;
-        Base::Vector3f base, axis;
+        Base::Vector3d base, axis;
         if (!getCylinder(radius, height, base, axis))
             return;
         base = getBasePoint(base + axis * height/2, axis, Location, Dist.getValue());
